### 2019-11-08 #1873 @whedon
[PRE REVIEW]: ScenTrees.jl: A Julia Library For Generating Scenario Trees and Scenario Lattices for Multistage Stochastic Optimization
**Submitting author:** @kirui93 (<a href="http://orcid.org/0000-0002-3679-4442">Kipngeno </a>)
**Repository:** <a href="https://github.com/kirui93/ScenTrees.jl" target ="_blank">https://github.com/kirui93/ScenTrees.jl</a>
**Version:** v0.1.2
