### 2019-11-11 c2a4344
Increase the size of the resulting lattice plot
