### 2019-11-13 #1873 @melissawm
Hello, @kirui93! There are some details on the paper:
- For the reference Høyland, K., & Wallace, S. W. (2001), there is a typo on the name of the journal (Management Science). Also, the DOI for this paper seems to be mistyped: I believe the correct number is 10.1287/mnsc.47.2.295.9834. Can you check this, please?
- For the reference Kovacevic, R. M., & Pichler, A. (2015) you should include a volume number (in this case, 235).
- For the reference Pflug, G. C., & Pichler, A. (2016), there is an extra subtitle that is not listed in the published version of the paper (here https://epubs.siam.org/doi/10.1137/15M1043376). Can you please check this?
- Finally, for some references there are Volume and Issue listed, and other not. Could you make sure that the references are standardized? Ideally, all references should have an issue number if applicable.

Thank you!
