+++ paper/paper.bib
@@ -5,8 +5,7 @@
   Year = {2001},
   Pages = {295--307},
   Volume = {47},
-  Comment = {cf. Klassen},
-  Doi = {10.1287/mnsc.47.4.295.9834}
+  Doi = {10.1287/mnsc.47.2.295.9834}
 }
 
 @Article{KovacevicPichler,
@@ -14,14 +13,10 @@
   title     = {Tree approximation for discrete time stochastic processes: a process distance approach},
   journal   = {Annals of Operations Research},
   year      = {2015},
+  volumne   = {235},
   pages     = {1--27},
   doi       = {10.1007/s10479-015-1994-2},
-  file      = {:KovacevicPichler.pdf:PDF},
-  groups    = {Pichler},
-  keywords  = {Stochastic processes and trees; Wasserstein and Kantorovich distance; Tree approximation; Optimal transport; Facility location; 90C15; 60B05; 90-08},
-  owner     = {aloisp},
-  publisher = {Springer US},
-  timestamp = {2015.09.09},
+  publisher = {Springer US}
 }
 
 @Article{Pflug2001,
@@ -31,11 +26,7 @@
   Year                     = {2001},
   Pages                    = {251--271},
   Volume                   = {89},
-  Doi                      = {10.1007/s101070000202},
-  File                     = {:Scenario Tree Generation, Pflug.pdf:PDF},
-  Keywords                 = {quantization error, process approximation},
-  Owner                    = {ap},
-  Timestamp                = {2011.04.01}
+  Doi                      = {10.1007/s101070000202}
 }
 
 @Article{Pflug2009,
@@ -45,11 +36,7 @@
   year =      {2009},
   volume =    {20},
   pages =     {1406--1420},
-  doi =       {10.1137/080718401},
-  file =      {:Pflug, Nested-Distribution, 2010.pdf:PDF},
-  keywords =  {process approximation, multistage, stochastic optimization},
-  owner =     {ap},
-  timestamp = {2009}
+  doi =       {10.1137/080718401}
 }
 
 @Article{PflugPichler2011,
@@ -60,11 +47,7 @@
   volume    = {22},
   number    = {1},
   pages     = {1--23},
-  doi       = {10.1137/110825054},
-  file      = {:NestedDistance.pdf:PDF},
-  groups    = {Pichler},
-  owner     = {ap},
-  timestamp = {2011.02.18},
+  doi       = {10.1137/110825054}
 }
 
 @Article{PflugPichlerDynScenarioGen,
@@ -75,11 +58,7 @@
   volume    = {62},
   number    = {3},
   pages     = {641--668},
-  doi       = {10.1007/s10589-015-9758-0},
-  file      = {:PflugPiSzenarioGen.pdf:PDF},
-  groups    = {Pichler},
-  owner     = {aloisp},
-  timestamp = {2015.04.21},
+  doi       = {10.1007/s10589-015-9758-0}
 }
 
 @Article{PflugPichler2016,
@@ -87,25 +66,16 @@
   title     = {From Empirical Observations to Tree Models for Stochastic Optimization: Convergence Properties},
   journal   = {{SIAM} Journal on Optimization},
   year      = {2016},
-  subtitle  = {Convergence of the Smoothed Empirical Process in Nested Distance},
   volume    = {26},
   number    = {3},
   pages     = {1715--1740},
-  doi       = {10.1137/15M1043376},
-  file      = {:TreesByKernels.pdf:PDF},
-  groups    = {Pichler},
-  owner     = {aloisp},
-  publisher = {Society for Industrial {\&} Applied Mathematics ({SIAM})},
[…]
