### 2019-11-13 #1873 @kirui93
Hello, @melissawm ,
Thank you so much for the corrections.
I think it was a typo for most of them but I have corrected them.

Many thanks.
