+++ paper/paper.bib
@@ -5,6 +5,7 @@
   Year = {2001},
   Pages = {295--307},
   Volume = {47},
+  number = {2},
   Doi = {10.1287/mnsc.47.2.295.9834}
 }
 
@@ -14,7 +15,8 @@
   journal   = {Annals of Operations Research},
   year      = {2015},
   volumne   = {235},
-  pages     = {1--27},
+  number    = {1},
+  pages     = {395--421},
   doi       = {10.1007/s10479-015-1994-2},
   publisher = {Springer US}
 }
@@ -26,6 +28,7 @@
   Year                     = {2001},
   Pages                    = {251--271},
   Volume                   = {89},
+  number                   = {2},
   Doi                      = {10.1007/s101070000202}
 }
 
@@ -35,6 +38,7 @@
   journal =   {SIAM Journal on Optimization},
   year =      {2009},
   volume =    {20},
+  number =    {3},
   pages =     {1406--1420},
   doi =       {10.1137/080718401}
 }
@@ -96,18 +100,18 @@
 		url={https://docs.julialang.org/en/v1}
 }
 
-@Article{Bagge2019,
-	title = {Parallel computing in Julia: Case study from Dept. Automatic Control, Lund University},
-	abstract = {This document outlines how to setup and run computations in parallel using Julia on a collectionof remote computers, such as computers in a university lab. After the environment has been setup,only minor modifications to serially executed code is necessary to enable parallel execution. Written for Julia version 1.0},
-	author = {Bagge, C. F.},
-	year = {2019},
-	month = {3},
-	day = {6},
-	language = {English},
-	series = {Technical reports TFRT-7657},
-	publisher = {Department of Automatic Control, Faculty of Engineering LTH, Lund University},
+% @Article{Bagge2019,
+% 	title = {Parallel computing in Julia: Case study from Dept. Automatic Control, Lund University},
+% 	abstract = {This document outlines how to setup and run computations in parallel using Julia on a collectionof remote computers, such as computers in a university lab. After the environment has been setup,only minor modifications to serially executed code is necessary to enable parallel execution. Written for Julia version 1.0},
+% 	author = {Bagge Carlson, Fredrik},
+% 	year = {2019},
+% 	month = {3},
+% 	day = {6},
+% 	language = {English},
+% 	series = {Technical reports TFRT-7657},
+% 	publisher = {Department of Automatic Control, Faculty of Engineering LTH, Lund University},
 	
-}
+% }
 
 
 @article{julia,
