+++ paper/paper.bib
@@ -92,26 +92,13 @@
 }
 
 @Manual{JULIA2019,
-		author={Julia},
-		year = {2019},
-		address={MIT},
-		organization={Julia Computing},
-		title={{Julia version 1.2}},
-		url={https://docs.julialang.org/en/v1}
+	author={Julia},
+	year = {2019},
+	address={MIT},
+	organization={Julia Computing},
+	title={{Julia v1.2}},
+	url={https://docs.julialang.org/en/v1}
 }
-
-% @Article{Bagge2019,
-% 	title = {Parallel computing in Julia: Case study from Dept. Automatic Control, Lund University},
-% 	abstract = {This document outlines how to setup and run computations in parallel using Julia on a collectionof remote computers, such as computers in a university lab. After the environment has been setup,only minor modifications to serially executed code is necessary to enable parallel execution. Written for Julia version 1.0},
-% 	author = {Bagge Carlson, Fredrik},
-% 	year = {2019},
-% 	month = {3},
-% 	day = {6},
-% 	language = {English},
-% 	series = {Technical reports TFRT-7657},
-% 	publisher = {Department of Automatic Control, Faculty of Engineering LTH, Lund University},
-	
-% }
 
 
 @article{julia,
