### 2019-11-19 #1873 @melissawm
@ChrisRackauckas would you be willing to review this paper for JOSS? Thank you!
