### 2019-11-19 #1873 @melissawm
@jarvist would you be willing to review this software paper for JOSS? Thank you!
