### 2019-11-19 #1873 @jarvist
> @jarvist would you be willing to review this software paper for JOSS? Thank you!

It's well outside my area of expertise, but I am happy to review as a naive user.
I have quite a few academic commitments at the moment, so it may take a while to fit this review in.
