### 2019-11-20 bf6b353
update lattice approxtn for larger node values
