### 2019-11-22 11fe52b
max and min for density plot
Change max and min for density plots.
As well the distance for resulting lattice
