### 2019-11-22 44644af
Remove the un-present figure
