### 2019-11-22 6326009
Use read directory instead of listing assets
