### 2019-11-22 0f04a5d
Remove assets
