### 2019-11-22 b21d669
compat versions for deps
