+++ Project.toml
@@ -1,31 +1,26 @@
 name = "ScenTrees"
 uuid = "4ec4db44-c288-11e9-3202-272e944f779a"
 authors = ["Kipngeno Kirui <kipngenokirui1993@gmail.com>"]
-version = "0.1.2"
+version = "0.1.3"
 
 [deps]
 CSV = "336ed68f-0bac-5ca0-87d4-7b16caf5d00b"
 DataFrames = "a93c6f00-e57d-5684-b7b6-d8193f3e46c0"
-Distributed = "8ba89e20-285c-5b6f-9357-94700520ee1b"
 Distributions = "31c24e10-a181-5473-b8eb-7969acd0382f"
 LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
 PyPlot = "d330b81b-6aea-500a-939a-2ce795aea3ee"
 Random = "9a3f8284-a2c9-5f02-9a11-845980a1fd5c"
-SharedArrays = "1a1011a3-84de-559e-8e89-a11a2f7dc383"
-SparseArrays = "2f01184e-e22b-5df5-ae63-d93ebab69eaf"
-StaticArrays = "90137ffa-7385-5640-81b9-e52037218182"
 Statistics = "10745b16-79ce-11e8-11f9-7d13ad32a3b2"
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
 XLSX = "fdbf4ff8-1666-58a4-91e7-1b58723a45e0"
 
 [compat]
-julia = "1"
-Distributions = "0.21, 1"
-StaticArrays = "^0.11.0, ^1"
-DataFrames = "^0.19.3, ^1"
-PyPlot = "^2.8.2, ^1"
-CSV = "^0.5.14, ^1"
-XLSX = "^0.5.8, ^1"
+julia = "1.0"
+Distributions = "0.21.8"
+DataFrames = "0.19.4"
+PyPlot = "2.8.2"
+CSV = "0.5.18"
+XLSX = "0.5.8"
 
 [extras]
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
