### 2019-11-29 0742105
update kde method
