### 2019-11-29 879d6c3
Minor final changes -Doc strings -Random RNG
