### 2019-12-03 962c0c2
Markov condition
