### 2019-12-05 5d7094f
Minor changes to initialization and plot
