### 2019-12-05 049d366
Update combat and increase version
