+++ Project.toml
@@ -1,7 +1,7 @@
 name = "ScenTrees"
 uuid = "4ec4db44-c288-11e9-3202-272e944f779a"
 authors = ["Kipngeno Kirui <kipngenokirui1993@gmail.com>"]
-version = "0.1.3"
+version = "0.1.4"
 
 [deps]
 CSV = "336ed68f-0bac-5ca0-87d4-7b16caf5d00b"
@@ -15,12 +15,12 @@ Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
 XLSX = "fdbf4ff8-1666-58a4-91e7-1b58723a45e0"
 
 [compat]
-julia = "1.0"
-Distributions = "0.21.8"
-DataFrames = "0.19.4"
+julia = "1"
+Distributions = "0.21.7,0.21.8,0.21.9,0.21.10"
+DataFrames = "0.18,0.19,0.20"
 PyPlot = "2.8.2"
-CSV = "0.5.18"
-XLSX = "0.5.8"
+CSV = "0.5.16,0.5.17,0.5.18"
+XLSX = "0.5.6,0.5.7,0.5.8"
 
 [extras]
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
