### 2019-12-06 73f2aab
Update plots and distance
