+++ Project.toml
@@ -1,7 +1,7 @@
 name = "ScenTrees"
 uuid = "4ec4db44-c288-11e9-3202-272e944f779a"
 authors = ["Kipngeno Kirui <kipngenokirui1993@gmail.com>"]
-version = "0.1.4"
+version = "0.1.5"
 
 [deps]
 CSV = "336ed68f-0bac-5ca0-87d4-7b16caf5d00b"
