+++ docs/src/tutorial/tutorial31.md
@@ -30,7 +30,7 @@ M_t = \max \{\sum_{t=1}^{t^\prime} \xi_t : t^\prime \leq t\} \ \text{with} \ \xi
 ```
 We insist that the user of this package _must write a function that takes no inputs_. This function should be, in a way, a hard-coded function that when you call, it produces an array with a certain user pre-determined characteristics.
 
-The above two processes in our package are in 4 stages and therefore can be used with a scenario tree or a scenario lattice with 4 stages. As already stated, the stages in a scenario tree or a scenario lattice starts from `0` to `T`. Each of these processes generates samples with a common starting point since a stochastic process has a deterministic starting value and the rest of the values are random.
+The above two processes in our package are in 4 stages and therefore can be used with a scenario tree or a scenario lattice with 4 stages. As already stated, the stages in a scenario tree or a scenario lattice starts from `1` to `T`. Each of these processes generates samples with a common starting point since a stochastic process has a deterministic starting value and the rest of the values are random.
 
 It is important to create a function that generates samples that are dependent on each other in `d` dimension. Most stochastic processes have states which depends on each other and so it is always important to consider dependency of the samples generated by the stochastic function.
 
