+++ docs/src/index.md
@@ -4,8 +4,7 @@ CurrentModule = ScenTrees
 
 # ScenTrees.jl
 
-!!! info
-	`ScenTrees.jl` is an actively developed package and new features and improvements are continually added.
+We present `Scentrees.jl` package for generating and improving scenario trees and scenario lattices for multistage stochastic optimization problems using stochastic optimization.
 
 ## Main features of the package
 
@@ -33,7 +32,7 @@ The following are the most important functions that ScenTrees.jl package provide
 
 1. `Process function`: This is a user-specified function that generates trajectories of a stochastic process that the user wants to approximate via a scenario tree or a scenario lattice. Scenario trees are used for discrete time stochastic processes while scenario lattices are natural discretization of the Markovian processes. This function should return an array in 2 dimension. It should not take any inputs else the user should consider using wrappers or function closures. This function depends on the number of stages in  the tree and the dimension of states of nodes in the tree. The user provides this function to the stochastic approximation process. Examples of inbuilt process functions are `GaussiansamplePath1D(),GaussianSamplePath2D(),RunningMaximum1D(),RunningMaximum2D(),path()`.
 
-2. `Tree`: This provides the structure of a scenario tree. It stores all the information of a tree. These information includes the name of the tree, the parent of each node, the state of each node and the probabilities from one node to another. An initial tree can be generated by `Tree([BStructure],dimenison)` where `Bstructure` is a branching vector and `dimension` is the dimension of states of nodes in the tree.
+2. `Tree`: This provides the structure of a scenario tree. It stores all the information of a tree. These information includes the name of the tree, the parent of each node, the state of each node and the probabilities from one node to another. An initial tree can be generated by `Tree([BStructure],dimenison)` where `BStructure` is a branching vector and `dimension` is the dimension of states of nodes in the tree.
 
 3. `Tree Approximation`: It returns a valuated probability scenario tree that approximates the stochastic process in the best possible way. The function `TreeApproximation!(Tree([BStructure],dimension),ProcessFunction,NoOfIterations,pNorm,rWasserstein)` takes:
 	* A scenario tree object from the `Tree()` function with a certain branching structure,
@@ -55,11 +54,3 @@ The scenario lattice approximation procedure follows the one for scenario tree a
 	* A data matrix in 2 dimension (i.e., `NxT` matrix where `N` is the number of trajectories and `T` is the number of stages.
 	* Kernel Distribution is the distribution that you want to use for the kernels. The default distribution is `Logistic distribution`. The user is free to choose any distribution he/she wishes to use based on the package `Distributions.jl`.
 	* The wrapper we use here states whether the trajectories that you want to create are either Markovian or not. When you set `Markovian=true`, the trajectories here are Markovian and so can be used to approximate a scenario lattice. Similarly, if you set `Markovian=false`, the trajectories are non-Markovian and therefore can be used to generate scenario trees.
-
-
-## Citing `ScenTrees.jl`
-
-For all the users of this package, we request that you cite the following paper:
-
-!!! info
-    To be added.

[…]
