### 2019-12-09 d0e5cdb
Clean the docs finally,.
