+++ docs/src/index.md
@@ -34,22 +34,11 @@ The following are the most important functions that ScenTrees.jl package provide
 
 2. `Tree`: This provides the structure of a scenario tree. It stores all the information of a tree. These information includes the name of the tree, the parent of each node, the state of each node and the probabilities from one node to another. An initial tree can be generated by `Tree([BStructure],dimenison)` where `BStructure` is a branching vector and `dimension` is the dimension of states of nodes in the tree.
 
-3. `Tree Approximation`: It returns a valuated probability scenario tree that approximates the stochastic process in the best possible way. The function `TreeApproximation!(Tree([BStructure],dimension),ProcessFunction,NoOfIterations,pNorm,rWasserstein)` takes:
-	* A scenario tree object from the `Tree()` function with a certain branching structure,
-	* A process function that we want to approximate,
-	* The number of iterations,
-	* Choice of the norm (Example: `max=0,sum=1,Euclidean=2 (default)`), and
-	* Wasserstein parameter (`rWasserstein = 2 (default)`)
+3. `Tree Approximation`: It returns a valuated probability scenario tree that approximates the stochastic process in the best possible way. The function `TreeApproximation!(Tree([BStructure],dimension),ProcessFunction,NoOfIterations,pNorm,rWasserstein)` takes: A scenario tree object from the `Tree()` function with a certain branching structure, a function generating trajectories, number of iterations, choice of the norm (Example: `max=0,sum=1,Euclidean=2 (default)`), and Wasserstein parameter (`rWasserstein = 2 (default)`). The number of iterations also represents the number of trajectories from the process function that we want to improve the scenario tree. For each iteration, we perform the stochastic approximation process which modifies one path in the tree towards the new trajectory and hence the approximating quality of the tree is improved in each iteration.
 
-The number of iterations also represents the number of trajectories from the process function that we want to improve the scenario tree. For each iteration, we perform the stochastic approximation process which modifies one path in the tree towards the new trajectory and hence the approximating quality of the tree is improved in each iteration.
+4. `Lattice Approximation`: It is used to approximate Markovian processes. Scenario lattices are natural discretization of Markov processes. The function `LatticeApproximation([BStructure],ProcessFunction,NoOfIterations)` takes: the branching structure of the lattice, function generating trajectories and the number of iterations. The scenario lattice approximation procedure follows the one for scenario tree approximation but here we find the closest lattice entry and use samples from the process function to improve it.
 
-4. `Lattice Approximation`: It is used to approximate Markovian processes. Scenario lattices are natural discretization of Markov processes. The function `LatticeApproximation([BStructure],ProcessFunction,NoOfIterations)` takes:
-	* The branching structure of the lattice,
-	* The process function that we want to approximate, and
-	* The number of iterations.
-The scenario lattice approximation procedure follows the one for scenario tree approximation but here we find the closest lattice entry and use samples from the process function to improve it.
-
-5. `Generation of scenarios from data`: This is a non-parametric technique for generating samples from a give data with unknown distribution. We use this process to generate new but different samples based of the data given using the kernel density estimation technique. The new samples can then be used to generate scenario trees and scenario lattices. The function `KernelScenarios(data,KernelDistribution;Markovian=true)` takes:
+5. `Generation of scenarios from data`: This is a non-parametric technique for generating samples from a give limited data with unknown distribution. We use this process to generate new but different samples based of the data given using the kernel density estimation technique. The new samples can then be used to generate scenario trees and scenario lattices. The function `KernelScenarios(data,KernelDistribution;Markovian=true)` takes:
 
 	* A data matrix in 2 dimension (i.e., `NxT` matrix where `N` is the number of trajectories and `T` is the number of stages.
[…]
