### 2019-12-12 #1912 @matbesancon
@melissawm a bunch of issues were opened, some being "nice-to-have" improvements on the package, but some being fairly necessary to navigate through the package and understand what is going on, how to use it & for which classes of problems. My opinion is that there is a fair bit of polishing still to do but I'll let the author voice in
