### 2019-12-12 #1912 @matbesancon
cc @kirui93
