### 2019-12-12 #1912 @kirui93
@matbesancon Thank you so much for mentioning that. I think I have implemented all the review comments that you had stated and that has enabled us to bump up the package. If you update to the new version, you will see that almost all the details are incorporated. What we are remaining with only is to change to use Plots.jl instead of PyPlot.jl which we will do it in some time to come as for now PyPlot.jl works well.
