### 2019-12-13 #1912 @kirui93
@matbesancon I also did polish the paper and I think it is much of straightforward and to the point now.
