### 2019-12-13 #1912 @kirui93
@matbesancon @juliohm I really believe that everything is now up to standard of the package we wanted to create. Your insightful review comments have been of great help.
