### 2019-12-13 #1912 @juliohm
@kirui93 please comment on each open issue to explain how you addressed them. There are many open issues still.
