### 2019-12-13 #1912 @kirui93
@juliohm I think most of the issues I have dealt with. So I will comment and close them if it is up to my side.
