### 2019-12-13 d986135
Docstrings and transportation parameter wrapped up.
