### 2019-12-16 #1912 @kirui93
@matbesancon @juliohm I did comment on the issues created and closed most of them as I had already dealt with them.
