### 2019-12-16 #1912 @juliohm
@kirui93 you didn't address many issues raised in the software paper for example. You also didn't mention how you addressed each issue on GitHub before closing them.

Please report on each issue above how you addressed the review comments, otherwise it is difficult to approve the paper for publication in JOSS.
