### 2019-12-16 #1912 @kirui93
> @kirui93 you didn't address many issues raised in the software paper for example. You also didn't mention how you addressed each issue on GitHub before closing them.
>
> Please report on each issue above how you addressed the review comments, otherwise it is difficult to approve the paper for publication in JOSS.

1.  General comments about the documentation - Initially, I had my docstrings placed in a wrong manner and with the help of @matbesancon I was able to rewrite the docstrings in the correct manner. And also @juliohm had mentioned briefly about the usage of "library" instead of "package" in most parts of the documentation, which I addressed accordingly.

2. @matbesancon indicated about the title and the project set up. He was suggesting to name the package from "ScenTrees.jl" to something like "ScenarioTrees.jl". Initially, we had a lot of suggestions for the name of the package  but we decided to go with "ScenTrees.jl" as it is unique and also in a way describes to our goal.

3. On naming and conventions, @matbesancon indicated about the function `Children{Vector{Int64}}` which looked like a constructor. As indicated, this function is an internal function of Tree and its main purpose is to reduce the number of times we are going to use the function in the stochastic approximation process. We access the children of the nodes each time in stochastic approximation iteration and hence saving the children as a vector at first with the "Chidren" function provides an easier and less expensive manner.

4. Another issue concerned incorporating AbstractTrees.jl. Though it was not a major issue, but as I had indicated, this package seems it is no longer maintained. And I couldn't find the documentation to detail its usage.

5. One of the most important issue that @matbesancon had stated was `every function which performs a random sampling should take in parameter the Random Number Generator (which is usually a subtype of Random.AbstractRNG)`. I found this as a good first issue as as he had indicated `making function runs deterministic` makes it really important. With this I used the package `Random.jl`. As stated in their documentation, `random number generation in Julia uses the Mersenne Twister library via MersenneTwister objects` and thus  I employed this for every function that involves random number generation.

6. There was quite a mislead regarding the Wasserstein distance and we thank @matbesancon for noting this out. In our code, we had used the term "WassersteinDistance" which was inappropriate. I stated in the issue that this was supposed to be the multistage distance and is duly corrected in the code. There is also the parameter `r` which is used for multistage distance and is also corrected well in all the codes.

7. The final issue that @juliohm opened was about improvements for the software paper. All the corrections he suggested are duly represented in the paper. The current version of the paper has all the corrections implemented.

Thank you so much @matbesancon and @juliohm for your patience as well as insightful comments about the software paper as well as the package in general.
