### 2019-12-17 #1912 @juliohm
@kirui93 can you please add these comments to the issues in your project repo instead? It is quite hard to follow any text here without context. Please understand that we have limited time to review.

I will try to read the issues again in the following weeks.

Thank you.
