### 2019-12-17 #1912 @kirui93
@juliohm @matbesancon Thank you so much for your patience. I have re-commented again on the important issues and re-opened them.
