### 2019-12-17 #1912 @kirui93
Some other issues are general comments that were just suggestions for improving the documentation as well as docstrings which I did implement immediately on statement.
