### 2019-12-19 1c1a471
Clean the docs and paper
