+++ paper/paper.md
@@ -27,75 +27,77 @@ bibliography: paper.bib
 
 # Summary
 
-In multistage stochastic optimization we are also interested in decision making under uncertainty. In general, stochastic processes have random and uncertain outcomes and decisions must be made at different stages of the process. Sometimes, it is generally intractable to solve mathematical programs with uncertain parameters described by an underlying probability distribution. For algorithmic treatment, such programs should be approximated by simpler ones, in the same manner as functions are represented as vectors on finite grids on digital computers. The approximation procedure is through discretization. The procedure of discretizing a stochastic process is called `scenario generation`. We depict the possible sequences of data for this processes in form of a `scenario tree` in the case of a discrete time stochastic process and a ``scenario lattice`` for Markov processes.
+Stochastic processes have random and uncertain outcomes and decisions must be made at different stages of the process. Mathematical programs with such uncertain parameters, described by an underlying probability distribution, are sometimes intractable to solve. For algorithmic treatment, such programs should be approximated into simpler ones, in the same manner as functions are represented as vectors on finite grids on digital computers. This approximation procedure is referred to as ``scenario generation``.
 
-The natural question when dealing with scenario generation for multistage stochastic optimization problems is how to approximate a continuous distribution of the stochastic process in such a way that the distance between the initial distribution and its approximation is minimized. We start with an initial scenario tree and use stochastic approximation to improve the values on the nodes of the tree. To quantify the quality of the approximation, we use process distance between the original process and the optimal scenario tree [@Pflug2009]. Process distance extends and generalizes the Wasserstein distance to stochastic processes. It was analyzed by @PflugPichler2011 and used by @KovacevicPichler directly to generate scenario trees.
+There is a vast literature describing different methods of scenario generation. Many authors including @Hoyland2001, @Pflug2001, @KovacevicPichler, @PflugPichlerDynScenarioGen and @PflugPichler2016 have addressed different approximation techniques for stochastic processes. But still there is no an open-source implementation of the various algorithms in the public domain.
 
-Generally, we consider a stochastic process $X$ over a discrete time, i.e., $X = (X_0,X_1,\ldots,X_T)$ where $X_0$ is a deterministic. A scenario tree is a discrete time and discrete state process approximating the process $X$. We represent a scenario tree by $\tilde{X} = (\tilde{X}_0,\tilde{X}_1,\ldots,\tilde{X}_T)$.
+We therefore present ``ScenTrees.jl``, a open-source Julia [@julia] package for generating scenario trees and scenario lattices which can be used, for example, for multistage stochastic optimization problems. It allows users to represent possible sequences of stochastic processes in form of a ``scenario tree`` in the case of a discrete time stochastic process and a ``scenario lattice`` for Markovian data processes. In extension, it also provides users with a platform for generating new and additional trajectories in case of a limited data using conditional density estimation. The theory and design of the ``ScenTrees.jl`` package follows the concept in @PflugPichlerDynScenarioGen directly. It starts with an initial tree, which is a qualified guess by expert opinion, and uses the stochastic approximation procedure to improve the values on the nodes of the tree with samples form the stochastic process to be approximated. The transition probabilities from one node to another are also addressed.
 
[…]
