### 2019-12-20 d857430
update plotting utilities
