### 2020-01-10 #1912 @kirui93
Hello @juliohm, I addressed all the issues that you had raised concerning the software paper. All the other issues also that were raised concerning the package are also addressed. You had mentioned some times back that I should comment on all the issues raised on how I addressed them and I did. Please have a look on my comments on the way I addressed the issues.
