### 2020-01-11 13c67a4
Update all functions
