+++ paper/paper.md
@@ -8,7 +8,7 @@ tags:
   - Stochastic Processes
   - Julia
 authors:
-  - name: Kipngeno Kirui
+  - name: Kipngeno Benard Kirui
     orcid: 0000-0002-3679-4442
     affiliation: "1"
   - name: Alois Pichler
@@ -45,7 +45,7 @@ To assess the quality of this approximation, a distance between the initial dist
 
 Implementation details and various examples to demonstrate different methods can be found in the package's documentation.^[Documentation: https://kirui93.github.io/ScenTrees.jl/latest]
 
-The following section provides a typical example on how to use ``ScenTrees.jl`` to generate scenario trees and scenario lattices by employing the stochastic approximation procedure and conditional density estimation.
+The following section provides a typical example on how to use ``ScenTrees.jl`` to generate scenario trees and scenario lattices by combining stochastic approximation procedure and conditional density estimation procedure.
 
 # Example: Scenario generation from observed trajectories
 
@@ -53,7 +53,7 @@ Consider the following comma-separated limited data. The data has $1000$ traject
 
 ```julia
 # Load the package
-julia> using ScenTrees  
+julia> using ScenTrees, CSV  
 # Load the data from a directory                 
 julia> df = CSV.read("../RData.csv");
 # Convert the DataFrame into a Matrix    
@@ -62,7 +62,7 @@ julia> data = Matrix(df);
 The following shows an example of a non-Markovian trajectory generated from the data using conditional density estimation by employing the Logistic distribution for the kernels.
 
 ```julia
-julia> Example = KernelScenarios(data, Logistic; Markovian=false)()
+julia> Example = KernelScenarios(data, Logistic; Markovian = false)()
 [1.5595,0.8150,1.5058,2.6475,4.6137]
 ```
 The generated data has a length equal to the number of columns of the original data. These generated trajectories are the ones we will use to approximate a scenario tree and a scenario lattice in the following sections.
@@ -73,7 +73,7 @@ We want to approximate the above data using a scenario tree with a branching vec
 
 ```julia
 julia> kernTree = TreeApproximation!(Tree([1,3,3,3,2],1),
-                  KernelScenarios(data, Logistic; Markovian=false),100000,2,2);
+                  KernelScenarios(data, Logistic; Markovian = false),100000,2,2);
 julia> treeplot(kernTree)
 julia> savefig("rwdataTree.pdf")
 ```
@@ -84,11 +84,11 @@ The number of possible trajectories in the scenario tree equals the number of le
 
 ## Approximation with a scenario lattice
 
-Consider a scenario lattice with a branching vector $(1,3,4,5,6)$. Clearly, this scenario lattice has $5$ stages as shown by the number of elements in the branching vector, which is equal to number of columns in the data. We consider $1,000,000$ iterations for the stochastic approximation algorithm. To generate this scenario lattice, we need Markovian samples and therefore we set `Markovian=true` to specify that the samples to be generated are Markovian, which is different for scenario trees.
+Consider a scenario lattice with a branching vector $(1,3,4,5,6)$. Clearly, this scenario lattice has $5$ stages as shown by the number of elements in the branching vector, which is equal to number of columns in the data. We consider $1,000,000$ iterations for the stochastic approximation algorithm and ``r=2`` parameter for the multistage distance. To generate this scenario lattice, we need Markovian trajectories and therefore we set `Markovian = true` to specify that the trajectories to be generated are Markovian, which is different for scenario trees.
 
 ```julia
 julia> rwdataLattice = LatticeApproximation([1,3,4,5,6],
-                       KernelScenarios(data, Logistic; Markovian=true), 1000000);
[…]
