+++ src/StochPaths.jl
@@ -1,7 +1,6 @@
 #All these examples of path are in 4 stages
 
-using Distributions
-using Random
+using Distributions, Random
 rng = MersenneTwister(01012019);
 
 """
@@ -10,7 +9,7 @@ rng = MersenneTwister(01012019);
 Returns a '4x1' dimensional array of Gaussian random walk
 """
 function GaussianSamplePath1D()
-    return vcat(0.0,cumsum(randn(rng,3,1),dims = 1)) #4 stages
+    return vcat(0.0, cumsum(randn(rng, 3, 1), dims = 1)) #4 stages
 end
 
 """
@@ -19,7 +18,7 @@ end
 Returns a '4x2' dimensional array of Gaussian random walk
 """
 function GaussianSamplePath2D()
-    gsmatrix = randn(rng,4,2) * [1.0 0.0; 0.9 0.3] #will create an (dimension x nstages) matrix
+    gsmatrix = randn(rng, 4, 2) * [1.0 0.0 ; 0.9 0.3] #will create an (dimension x nstages) matrix
     gsmatrix[1,:] .= 0.0
     return cumsum(gsmatrix .+ [1.0 0.0], dims = 1)
 end
@@ -30,8 +29,8 @@ end
 Returns a '4x1' dimensional array of Running Maximum process.
 """
 function RunningMaximum1D()
-    rmatrix = vcat(0.0,cumsum(randn(rng,3,1),dims = 1))
-    for i = 2:4
+    rmatrix = vcat(0.0, cumsum(randn(rng, 3, 1), dims = 1))
+    for i = 2 : 4
         rmatrix[i] = max.(rmatrix[i-1], rmatrix[i])
     end
     return rmatrix
@@ -43,12 +42,12 @@ end
 Returns a '4x2' dimensional array of Running Maximum process.
 """
 function RunningMaximum2D()
-    rmatrix = vcat(0.0,cumsum(randn(rng,3,1),dims = 1))
-    rmatrix2D = zeros(4,2)
+    rmatrix = vcat(0.0, cumsum(randn(rng, 3, 1), dims = 1))
+    rmatrix2D = zeros(4, 2)
     rmatrix2D[:,1] .= vec(rmatrix)
-    for j=2:2
-        for i=2:4
-            rmatrix2D[i,j] = max.(rmatrix[i-1],rmatrix[i])
+    for j = 2 : 2
+        for i = 2 : 4
+            rmatrix2D[i,j] = max.(rmatrix[i-1], rmatrix[i])
         end
     end
     return rmatrix2D * [1.0 0.0; 0.9 0.3]
@@ -57,8 +56,8 @@ end
 """
 	path()
 
-Returns a sample of prices following the Geometric random process.
+Returns a sample of stock prices following the a simple random random process.
 """
 function path()
-    return  100 .+ 50 * vcat(0.0,cumsum(randn(rng,3,1),dims = 1))
+    return  100 .+ 50 * vcat(0.0, cumsum(randn(rng, 3, 1), dims = 1))
 end
