+++ src/TreeStructure.jl
@@ -1,13 +1,9 @@
 
-using Distributions
-using Statistics
-using PyPlot
-using Random
+using Distributions, Statistics, PyPlot, Random
 rng = MersenneTwister(01012019);
 
 #This defines the tree structure.
-
-mutable struct Tree                                                                               #Tree structure
+mutable struct Tree
     name::String                        # name of the tree
     parent::Vector{Int64}               # parents of nodes in the tree
     children::Vector{Vector{Int64}}     # successor nodes of each parent
@@ -22,29 +18,29 @@ mutable struct Tree                                                             
     function Children(parent::Vector{Int64})
         allchildren = Vector{Vector{Int64}}([])
         for node in unique(parent)
-            push!(allchildren,[i for i = 1:length(parent) if parent[i] == node])
+            push!(allchildren, [i for i = 1 : length(parent) if parent[i] == node])
         end
         return allchildren
     end
     Tree(name::String,parent::Vector{Int64},
     children::Vector{Vector{Int64}},
-    state::Matrix{Float64},probability::Matrix{Float64})=new(name,parent,Children(parent),state,probability)
+    state::Matrix{Float64}, probability::Matrix{Float64})=new(name, parent, Children(parent), state, probability)
 
     """
     	Tree(identifier::Int64)
 
     Returns some examples of predefined trees.
-    These are (0,302,303,304,305,306,307,402,404,405).
+    These are (0, 302, 303, 304, 305, 306, 307, 402, 404, 405).
     You can call any of the above tree and plot to see the properties of the tree.
     """
     function Tree(identifier::Int64)
         self= new()
         if identifier==0
             self.name = "Empty Tree"
-            self.parent = zeros(Int64,0)
+            self.parent = zeros(Int64 , 0)
             self.children = Children(self.parent)
-            self.state = zeros(Float64,0)'
-            self.probability = ones(Float64,0)'
+            self.state = zeros(Float64, 0)'
+            self.probability = ones(Float64, 0)'
         elseif identifier == 302
             self.name = "Tree 1x2x2"
             self.parent =[0,1,1,2,2,3,3]
@@ -91,14 +87,12 @@ mutable struct Tree                                                             
             self.name = "Tree 1x2x2x2"
             self.parent = [0,1,1,2,2,3,3,4,4,5,5,6,6,7,7]
             self.children = Children(self.parent)
-            #self.state= [10.0 11.0 8.0 12.0 9.0 6.0 10.0 13.0 10.01 8.01 12.01 9.01 4.0 11.0 12.99]'
             self.state = [10.0 12.0 8.0 15.0 11.0 9.0 5.0 18.0 16.0 13.0 11.0 10.0 7.0 6.0 3.0]'
             self.probability = [1.0 0.8 0.7 0.3 0.2 0.8 0.4 0.6 0.2 0.5 0.5 0.4 0.6 0.7 0.3]'
         elseif identifier == 4022
             self.name = "Tree 1x2x2x2"
             self.parent = [0,1,1,2,2,3,3,4,4,5,5,6,6,7,7]
             self.children = Children(self.parent)
-            #self.state= [10.0 11.0 8.0 12.0 9.0 6.0 10.0 13.0 10.01 8.01 12.01 9.01 4.0 11.0 12.99]'
[…]

[…]
