### 2020-01-22 c6c0a55
Test for 1.3 and paper
