+++ .travis.yml
@@ -2,13 +2,13 @@
 language: julia
 os:
   - linux
-  - osx
+#  - osx
   
 julia:
   - 1.0
   - 1.1
   - 1.2
-  - nightly
+  - 1.3
   
 branches:
  only:
