### 2020-01-29 kirui93/ScenTrees.jl#15 @matbesancon
treeplot function
```julia
function treeplot(trr::Tree, fig= 1)
    if !isempty(fig)
        figure(fig)
    end
    # ...
```

it is not trivial what exactly users should pass as `fig` argument. `isempty` for a number is always false. What else could they pass to have it empty? In that case what happens?
