### 2020-01-29 kirui93/ScenTrees.jl#15 @matbesancon
Linking to https://github.com/openjournals/joss-reviews/issues/1912
