+++ .travis.yml
@@ -9,17 +9,17 @@ julia:
   - 1.1
   - 1.2
   - 1.3
+  - nightly
+
+after_success:
+  - julia -e 'import Pkg; Pkg.add("Coverage"); using Coverage; Codecov.submit(process_folder()); Coveralls.submit(process_folder())'
   
 branches:
  only:
    - master
-   
-matrix:
-  fast_finish: true
-  allow_failures:
-    - julia: nightly
-   
-codecov: true
+   - /^v\d+\.\d+(\.\d+)?(-\S*)?$/
+ 
+#codecov: true
 
 notifications:
   email: false
@@ -27,19 +27,20 @@ notifications:
 env:
   - PYTHON=""
 
-#script:
-#  - julia -e 'using Pkg; Pkg.test("ScenTrees",coverage=true)'
 
 jobs:
+  fast_finish: true
+  allow_failures:
+    - julia: nightly
   include:
     - stage: "Documentation"
-      julia: 1.1
+      julia: 1.0
       os: linux
       script:
         - julia --project=docs/ -e 'using Pkg; Pkg.develop(PackageSpec(path=pwd())); Pkg.instantiate()'
         - julia --project=docs/ -e 'using Pkg; Pkg.add("Documenter"); using Documenter; include("docs/make.jl")'
       after_success: skip
        
-after_success:
-  - julia -e 'using Pkg; cd(Pkg.dir("ScenTrees")); Pkg.add("Coverage"); using Coverage; Codecov.submit(Codecov.process_folder())'
-  - julia -e 'using Pkg; cd(Pkg.dir("ScenTrees")); Pkg.add("Coverage"); using Coverage; Coveralls.submit(Coveralls.process_folder())'
+#after_success:
+#  - julia -e 'using Pkg; cd(Pkg.dir("ScenTrees")); Pkg.add("Coverage"); using Coverage; Codecov.submit(Codecov.process_folder())'
+#  - julia -e 'using Pkg; cd(Pkg.dir("ScenTrees")); Pkg.add("Coverage"); using Coverage; Coveralls.submit(Coveralls.process_folder())'
