### 2020-01-29 kirui93/ScenTrees.jl#15 @kirui93
Sometimes, a user may want to have his/her picture larger than what the package returns. In this case, having `fig = (8,6)` for example will increase the size of the image. That is why we have that optional argument.
