### 2020-01-29 72548a7
snake case method for functions
