+++ src/LatticeApprox.jl
@@ -9,7 +9,7 @@ mutable struct Lattice
 end
 
 """
-	LatticeApproximation(bstructure::Array{Int64,1}, path::Function, nIterations::Int64, r::Int64 = 2)
+	lattice_approximation(bstructure::Array{Int64,1}, path::Function, nIterations::Int64, r::Int64 = 2)
 
 Returns an approximated lattice approximating the stochastic process provided.
 
@@ -19,7 +19,7 @@ path - function generating samples from known distribution with length equal to 
 nIterations - number of iterations for stochastic approximation algorithm.
 r - parameter for the transportation distance.
 """
-function LatticeApproximation(bstructure::Array{Int64,1}, path::Function, nIterations::Int64, r::Int64 = 2)
+function lattice_approximation(bstructure::Array{Int64,1}, path::Function, nIterations::Int64, r::Int64 = 2)
     t_dist = 0.0                               # multistage distance
     T = length(bstructure)     # number of stages in the scenario lattice
     states = [zeros(bstructure[j], 1) for j = 1 : T]   # States of the lattice at each time t
@@ -53,7 +53,7 @@ end
 
 Returns a plot of a lattice.
 """
-function PlotLattice(lt::Lattice,fig = 1)
+function plot_lattice(lt::Lattice,fig = 1)
     if !isempty(fig)
         figure(figsize=(6,4))
     end
