+++ docs/make.jl
@@ -2,16 +2,13 @@ using Documenter, ScenTrees
 
 #const ASSETS = readdir(joinpath(@__DIR__, "src", "assets"))
 
-#isCI = get(ENV, "CI", nothing) == "true" #Travis populates this env variable by default
-
 makedocs(
+	modules = [ScenTrees],
 	sitename =  "ScenTrees.jl",
 	authors = "Kipngeno Kirui",
 	clean = true,
 	doctest = false,
-	format = Documenter.HTML(
-			prettyurls = get(ENV, "CI", nothing) == "true"
-	),
+	format = Documenter.HTML(prettyurls = get(ENV, "CI", nothing) == "true"),
 	pages = ["Home" => "index.md",
 		"Tutorials" => Any["tutorial/tutorial1.md",
 				    "tutorial/tutorial2.md",
@@ -24,9 +21,8 @@ makedocs(
 		]
 )
 
-#if isCI
 deploydocs(
 	repo = "github.com/kirui93/ScenTrees.jl.git",
+	target = "build",
 	versions = ["stable" => "v^", "v#.#", "dev" => "master"]
 )
-#end
