+++ docs/make.jl
@@ -3,7 +3,6 @@ using Documenter, ScenTrees
 #const ASSETS = readdir(joinpath(@__DIR__, "src", "assets"))
 
 makedocs(
-	modules = [ScenTrees],
 	sitename =  "ScenTrees.jl",
 	authors = "Kipngeno Kirui",
 	clean = true,
@@ -24,5 +23,7 @@ makedocs(
 deploydocs(
 	repo = "github.com/kirui93/ScenTrees.jl.git",
 	target = "build",
-	versions = ["stable" => "v^", "v#.#", "dev" => "master"]
+	deps = nothing,
+	make = nothing
+	#versions = ["stable" => "v^", "v#.#", "dev" => "master"]
 )
