### 2020-02-02 kirui93/ScenTrees.jl#15 @matbesancon
ok I see, this should be documented in the docstring of the function
