### 2020-02-02 #1912 @matbesancon
As an update, only https://github.com/kirui93/ScenTrees.jl/issues/15 remains on the documentation side, I'll do a last pass on the paper but it starts to all look good to me, @juliohm do you have something more?
