### 2020-02-02 kirui93/ScenTrees.jl#15 @kirui93
I will add it in the docstrings as I update the stable package version.
