### 2020-02-02 c3d8de5
docstrings for tree_plot function
