### 2020-02-02 a534aac
Update link to stable documentation
