### 2020-02-02 kirui93/ScenTrees.jl#15 @kirui93
@matbesancon  It is done. I have also updated all the stable documentation in the README and the paper. As well I have bumped up the version to v0.20.
