### 2020-02-03 #1912 @matbesancon
@kirui93 can you check the DOI above, there is one that needs removing the https://doi.org prefix
