+++ paper/paper.bib
@@ -132,6 +132,6 @@
   volume = {9},
   number = {2},
   pages = {129--132},
-  doi = {https://doi.org/10.1016/0167-7152(92)90006-Q},
+  doi = {10.1016/0167-7152(92)90006-Q},
   publisher ={Elsevier}
 }
