### 2020-02-03 #1912 @kirui93
Yeah. It was a slight mistake in the .bib file. I have removed it.
