### 2020-02-03 691dc92
Remove invalid DOI
