+++ paper/paper.bib
@@ -123,15 +123,3 @@
   pages     = {1156--1168},
   doi       = {10.1016/j.ejor.2016.11.028}
 }
-
-@Article{Jones1990,
-  author = {M. C. Jones},
-  title = {The performance of kernel density functions in kernel distribution function},
-  journal = {Statistics and Probability Letters},
-  year = {1990},
-  volume = {9},
-  number = {2},
-  pages = {129--132},
-  doi = {10.1016/0167-7152(92)90006-Q},
-  publisher ={Elsevier}
-}
