### 2020-02-03 #1912 @kirui93
I have removed this reference in `paper.pip` file because I didn't even cite it. I am wondering why whedon is saying invalid to a reference already removed
