### 2020-02-06 #1912 @melissawm
@kirui93 before we finish, can I ask you to look into some details?
- In the first paragraph of the summary, the word "generation" in "scenario generation" is broken; can you fix this?
- In the reference Høyland, K., & Wallace, S. W. (2001) there is a typo in the word Science in "Management Sceince"
- In the reference Kovacevic, R. M., & Pichler, A. (2015), I believe you're missing a 235 as the citation seems to be "Kovacevic, R.M., Pichler, A. Tree approximation for discrete time stochastic processes: a process distance approach. Ann Oper Res 235, 395–421 (2015). https://doi.org/10.1007/s10479-015-1994-2"

After those I'll ask you to generate a new proof and double check so we can generate the final archive. Thanks!
