+++ paper/paper.bib
@@ -1,7 +1,7 @@
 @Article{Hoyland2001,
   Title = {Generating Scenario Trees for multistage decision problems},
-  AUthor = {Kjetil H{\o}yland and Stein William Wallace},
-  Journal = {Management Sceince},
+  Author = {Kjetil H{\o}yland and Stein William Wallace},
+  Journal = {Management Science},
   Year = {2001},
   Pages = {295--307},
   Volume = {47},
@@ -14,7 +14,7 @@
   title     = {Tree approximation for discrete time stochastic processes: a process distance approach},
   journal   = {Annals of Operations Research},
   year      = {2015},
-  volumne   = {235},
+  volume   = {235},
   number    = {1},
   pages     = {395--421},
   doi       = {10.1007/s10479-015-1994-2},
