### 2020-02-06 #1912 @kirui93
@melissawm Thanks for the comments.
* Is there a way to prevent a word break really in a markdown document?
