### 2020-02-18 #1912 @kirui93
Hello @melissawm

The DOI for the Zenodo archive is

https://doi.org/10.5281/zenodo.3672205

The version adapted to the DOI is the latest version v0.2.2

Thank you.
