### 2020-02-18 #1912 @melissawm
Hi @kirui93 please check the metadata on the Zenodo archive, especially the authors list, as the citation reads "Kipngeno Kirui, Júlio Hoffimann, & Julia TagBot. (2020, February 18). kirui93/ScenTrees.jl: Release for zenodo (Version v0.2.2). Zenodo. http://doi.org/10.5281/zenodo.3672205"
