### 2020-02-18 #1912 @kirui93
Hello @melissawm,
 At first I had confused how to upload to Zenodo. Here is the right DOI

http://doi.org/10.5281/zenodo.3673939

The version adapted is still v0.2.2

Thank you for your patience.
