### 2020-02-18 #1912 @melissawm
@openjournals/joss-eics I think this is ready!

Thanks @kirui93 and reviewers @matbesancon and @juliohm !
